+++ src/yolanda.h
@@ -1,2 +1,3 @@
 #pragma once
 
+#include "yo/lambda.h"
